GoGeanAI-project---01--main/GoGeanAI-project---01--main/gogenAi/linkedin_project/requirements.txt
#os
time
random
pickle
sys
gspread
openai
logging
datetime
re
math
python-dotenv
cryptography
selenium
multiprocessing
oauth2client
timezonefinder
geopy
zoneinfo   
mysql.connector